Move tests/oracles.py -> tests/common/oracles.py
