+++ tests/common/oracles.py
